fix: everything broke!
Accidentally used an assignment operator `=` instead of a comparison
`===`. Should have double checked before pushing to prod, whoops!

Also, having an invalid input and then clicking "cancel" no longer locks
the player in the game - playerInput is allowed to drop out of scope on
each loop of the function to prevent it from remembering a bugged value